Fixing bugs in JoinOperator.
The previous implementation used per-state watermark callbacks (that had
mutable access to the stream state) to clean up the state after the join
callback is invoked. However, this was buggy, since consecutive
watermark callbacks could clear up state that might be required for the
join callback (when one stream is lagging behind the other). This fix
cleans up state after the join callback is called.